requirements.txt
Flask
opencv-python
numpy
requests
python-dotenv